+++ crates/client/db/src/storage_handler/mod.rs
@@ -10,8 +10,6 @@ use starknet_api::hash::StarkFelt;
 use starknet_api::state::StorageKey;
 use starknet_types_core::felt::Felt;
 
-use self::block_hash::BlockHashView;
-use self::block_number::BlockNumberView;
 use self::block_state_diff::BlockStateDiffView;
 use self::class_trie::{ClassTrieView, ClassTrieViewMut};
 use self::contract_class_data::{ContractClassDataView, ContractClassDataViewMut};
@@ -24,8 +22,6 @@ use self::history::HistoryError;
 use crate::DeoxysBackend;
 
 pub mod benchmark;
-pub mod block_hash;
-pub mod block_number;
 pub mod block_state_diff;
 mod class_trie;
 mod codec;
@@ -255,14 +251,6 @@ pub fn contract_nonces() -> ContractNoncesView {
 
 pub fn contract_nonces_mut() -> ContractNoncesViewMut {
     ContractNoncesViewMut::new()
-}
-
-pub fn block_number() -> BlockNumberView {
-    BlockNumberView
-}
-
-pub fn block_hash() -> BlockHashView {
-    BlockHashView
 }
 
 pub fn block_state_diff() -> BlockStateDiffView {